Move DefaultOptions to dps package
