+++ cmd/extract-block-events/main.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/dgraph-io/badger/v2"
 	"github.com/fxamacker/cbor/v2"
-	"github.com/optakt/flow-dps/service/chain"
+	"github.com/optakt/flow-dps/models/dps"
 	"github.com/rs/zerolog"
 	"github.com/spf13/pflag"
 
@@ -68,7 +68,7 @@ func main() {
 	log = log.Level(level)
 
 	// Initialize the protocol state database we will use.
-	opts := chain.DefaultOptions(flagData).WithLogger(nil)
+	opts := dps.DefaultOptions(flagData).WithLogger(nil)
 	db, err := badger.Open(opts)
 	if err != nil {
 		log.Fatal().Err(err).Msg("could not open database")
