+++ service/chain/protocol_state.go
@@ -19,8 +19,6 @@ import (
 	"fmt"
 
 	"github.com/dgraph-io/badger/v2"
-	"github.com/dgraph-io/badger/v2/options"
-
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/storage"
 	"github.com/onflow/flow-go/storage/badger/operation"
@@ -30,23 +28,6 @@ import (
 
 type ProtocolState struct {
 	db *badger.DB
-}
-
-func DefaultOptions(dir string) badger.Options {
-	return badger.DefaultOptions(dir).
-		WithMaxTableSize(256 << 20).
-		WithValueLogFileSize(64 << 20).
-		WithTableLoadingMode(options.FileIO).
-		WithValueLogLoadingMode(options.FileIO).
-		WithNumMemtables(1).
-		WithKeepL0InMemory(false).
-		WithCompactL0OnClose(false).
-		WithNumLevelZeroTables(1).
-		WithNumLevelZeroTablesStall(2).
-		WithLoadBloomsOnOpen(false).
-		WithIndexCacheSize(2000 << 20).
-		WithBlockCacheSize(0).
-		WithLogger(nil)
 }
 
 func FromProtocolState(db *badger.DB) *ProtocolState {
